# laravel/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Category CRUD",
        "description": "Category CRUD example",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/product?page={page}&title={title}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "02c8f1e6165fba91c6c88253fa7b6ea5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by title",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Products."
                    }
                }
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "256b169474326e6a0209f6e8ba1ad056",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation has been fault"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "54c88f9e21a81007799d97a52d5c84ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One category found."
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "400": {
                        "description": "Request validation fault"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "535b86989adaa5c19491861009675350",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation has fault"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "81469953ef762653a2c350d45421ad27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "id not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}